requirements.txt
simpleITK==2.1.0
torchsummaryX
nibabel
pillow 
tensorboard
gdown
pytorch-ignite==0.4.4
itk
tqdm
lmdb
psutil
pandas
einops
scikit-image
monai >= 0.8.1
